refactor(settings): rename VisualEffectsMenu directory to AppSettingsMenu (#1150)
